\name{KENDL-package}
\alias{KENDL-package}
\alias{KENDL}
\docType{package}
\title{
Kernel-smoothed Nonparametric Methods for Environmental Exposure Data Subject to Detection Limits
}
\description{
Calculate the kernel-smoothed nonparametric estimator for the exposure distribution in presence of detection limits.

}
\details{
\tabular{ll}{
Package: \tab KENDL\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-03-13\cr
License: \tab  GPL-2\cr
}

}
\author{
Yuchen Yang <yuchen.y@uky.edu>, Li Chen <lichenuky@uky.edu>

Maintainer: Yuchen Yang <yuchen.y@uky.edu>

}

