\name{mypackage-internal}
        \alias{KRKM.default}
        \alias{ker}
        \alias{silverman}
        \title{Internal mypackage Functions}
        \description{
          Internal mypackage functions
        }
        \details{
          These are not to be called by the user.
        }
        \keyword{ internal }
